+++ app.js
@@ -11,6 +11,7 @@ const path = require("path");
 
 const users = require("./routes/api/users");
 const posts = require("./routes/api/posts");
+const comments = require("./routes/api/comments");
 
 //below for heroku
 // app.use("/", express.static(path.join(__dirname, "/client/build")));
@@ -34,22 +35,20 @@ app.get("/logout", function (req, res) {
   res.redirect("/");
 });
 
-
 app.use(passport.initialize());
 require("./config/passport")(passport);
-
 
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(bodyParser.json());
 
 app.use("/api/users", users);
-app.use("/api/posts", posts)
+app.use("/api/posts", posts);
 app.use("/api/document", fileUploadRoutes);
+app.use("/api/comments", comments);
 
 
 const port = process.env.PORT || 5000;
 app.listen(port, () => console.log(`Server is running on port ${port}`));
-
 
 //below for heroku
 // app.get("*", (req, res) => {
