+++ bin/www
@@ -1,41 +1,33 @@
-#!/usr/bin/env node
-
 require('dotenv').config();
 var app = require('../app');
 var debug = require('debug')('mean-app:server');
 var http = require('http');
 
-// Get port from environment and store in Express.
 var port = normalizePort(process.env.PORT || '3000');
 app.set('port', port);
 
 
-// Create HTTP server.
 var server = http.createServer(app);
 
-// Listen on provided port, on all network interfaces.
+
 server.listen(port);
 server.on('error', onError);
 server.on('listening', onListening);
 
-// Normalize a port into a number, string, or false.
 function normalizePort(val) {
   var port = parseInt(val, 10);
 
   if (isNaN(port)) {
-    // named pipe
     return val;
   }
 
   if (port >= 0) {
-    // port number
     return port;
   }
   return false;
 }
 
 
-// Event listener for HTTP server "error" event.
 function onError(error) {
   if (error.syscall !== 'listen') {
     throw error;
@@ -45,7 +37,6 @@ function onError(error) {
     ? 'Pipe ' + port
     : 'Port ' + port;
 
-  // handle specific listen errors with friendly messages
   switch (error.code) {
     case 'EACCES':
       console.error(bind + ' requires elevated privileges');
@@ -60,8 +51,6 @@ function onError(error) {
   }
 }
 
-
-// Event listener for HTTP server "listening" event.
 function onListening() {
   var addr = server.address();
   var bind = typeof addr === 'string'
