+++ models/Comment.js
@@ -1,11 +1,8 @@
+
 const mongoose = require("mongoose");
 const Schema = mongoose.Schema;
 
 const CommentSchema = new Schema({
-    user: {
-        type: Schema.Types.ObjectId,
-        ref: "users",
-    },
     post: {
         type: Schema.Types.ObjectId,
         ref: "posts",
@@ -20,4 +17,4 @@ const CommentSchema = new Schema({
     },
 });
 
-module.exports = Comment = mongoose.model("comment", CommentSchema);
+module.exports = Comment = mongoose.model("comments", CommentSchema);
